requirements.txt
asgiref==3.10.0
Django==5.2.7
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
tzdata==2025.2
django-extensions
pyOpenSSL
werkzeug
cryptography
django-cleanup

# --- BERSERK ARC ADDITIONS ---
Faker==25.8.0
django-filter==24.2